Update and rename paddle-adventure.tsx to guided-paddle-adventure.tsx
