+++ src/containers/jp/class-signup/types/other/guided-paddle-adventure.tsx
@@ -1,0 +1,16 @@
+import * as React from "react";
+import { ClassType } from "../../class-description";
+
+const data: ClassType = {
+	typeId: 1241,
+	typeName: "Guided Paddle Adventure",
+	prereq: "None",
+	sessionCt: 1,
+	sessionLength: 1.5,
+	classSize: 20,
+	description: (<React.Fragment>
+		Guided Paddle Adventure takes the Kayak rating received in Beginner Sailing to another level. In class, with instructor supervision and organization, Juniors may paddle to new and distant places not normally allowed or to join in organized kayak games and relays! They might take a trip up the lagoon, play Capture the Life-Jacket or Ultimate Kayak, go on a scavenger hunt, or run a relay race in the mooring field. Juniors who have their Stand-Up Paddleboard ratings can take out Stand-Up Paddleboards instead of kayaks.
+	</React.Fragment>)
+}
+
+export default data;
